mappers & favs update
mappers converted to extension,
favorites page added

+++ app/src/main/java/com/softcross/onepiece/core/common/extension/mapExtensions/CrewMapExtension.kt
@@ -42,7 +42,7 @@ fun FavoriteCrewEntity.toCrew(): Crew {
 
 fun Crew.toFavoriteCrew(): FavoriteCrewEntity {
     return FavoriteCrewEntity(
-        id = id.toInt(),
+        id = id,
         name = crewName,
         totalBounty = crewTotalBounty,
         mainShip = crewMainShip,
